+++ api/routes/reservations.route.js
@@ -1,9 +1,11 @@
 
-const getReservations=require('../controllers/getReservations');
+const {getReservations,postReservation} = require('../controllers/controllerReservations');
 const express =require('express')
 const routerReservation=express.Router()
 
-routerReservation.get('/', getReservations)
+routerReservation
+    .get('/', getReservations)
+    .post('/', postReservation);
 
- module.exports=routerReservation;
+module.exports=routerReservation;
